Update README.md
Human activity detection is very important in understanding the various types of activities performed by a human based on the movement of his hands, legs and other useful information for prediction. 
With the advent of sensors and other gauging tools, it is easy to track some of the intricate information related to the human. Therefore, it is also important to classify the type of action being performed by the user using this intricate information. 
The deep learning algorithms would classify the type of action performed by the user by taking into consideration some of the input features such as hand movements, heart rate and so on. 
Therefore, Keras is used as a tool for performing the deep learning tasks for prediction. In addition, other machine learning models such as linear SVC has been used in the process of the prediction of human activity. 
Finally, all the models have been compared using prettytable and listed showcasing the accuracy and performance of the models.